+++ packages/winnow/src/geometry/index.js
@@ -1,5 +1,3 @@
-const Terraformer = require('terraformer')
-
 function createClause (options) {
   if (!options.geometry) return
   const spatialPredicate = options.spatialPredicate || 'ST_Intersects'