+++ packages/winnow/src/select/toEsri.js
@@ -1,0 +1,4 @@
+module.exports = function (projection) {
+  if (projection) return 'SELECT properties as attributes, esriGeom(project(geometry, ?)) as geometry FROM ?'
+  else return 'SELECT properties as attributes, esriGeom(geometry) as geometry FROM ?'
+}